+++ employee/urls.py
@@ -9,6 +9,7 @@ urlpatterns = [
     path("password/change", views.changePassword.as_view(), name="change_password"),
     path("office_hours", views.officeHours.as_view(), name="office_hours"),
     path("form/<form_id>/upload", views.uploadForm.as_view(), name="upload_form"),
+    path("form/<form_id>/sign", views.automaticallySignForm.as_view(), name="sign_form"),
     path(
         "mass-assign-paperwork/<selected>",
         views.massAssignPaperwork.as_view(),